test: add LBA likelihood
